[arc.SurfaceOrientation] Poll if color quadrants matched up
The SurfaceOrientation test is failing on many devices because the
screenshot that is used to compare to expected colors is taken too soon.
The screenshot shows a gray window or blurry colored blocks that are
fading into the front.

Change the test to poll for 10 seconds and keep trying to get a
screenshot that has the right color quadrants.

BUG=b:157503221
TEST=Pass on kohaku DUT
TEST=Test on leased magpie and blipper devices which were two devices
that were failing in stainless and appeared to fail alot after leasing
until this change.

Change-Id: I033d33a18e5b4e38a5e754a7a7676877be766138