[arc.SurfaceOrientation] Pause before screenshot (1 second)
The SurfaceOrientation test is failing on many devices because the
screenshot that is used to compare to expected colors is taken too soon.
The screenshot shows a gray window or blurry colored blocks that are
fading into the front.

Add a 1 second timeout for now that will hopefully fix this flakiness.
It may need to be increased in the future.

BUG=b:157503221
TEST=Pass on kohaku DUT
TEST=Observe stainless results on all other devices for informational test

Change-Id: I033d33a18e5b4e38a5e754a7a7676877be766138