tast-tests: Add timeout param to arc.New
Every other test that uses arc.New either opts-in first or uses
arc.Enabled() precondition which opts-in to ARC therefore ARC has
already started at that point.

enterprise.ARCProvisioning test relies on policy to start ARC and test
often fails because ARC hasn't finished booting within the default 2
minute timeout.

In this change, we're adding a timeout parameter to arc.New so we can
give a bigger timeout value when using arc.New in this test.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: Ief4834899245b68dd88aabce49fee67d3474d321