arcappcompat: Fix tablet mode failures
Fixed an issue where after the Full Screen tablet test was run, calling
EnsureTabletModeEnabled in an attempt to switch to tablet mode would
cause the test to error out. This is resolved by only calling
EnsureTabletModeEnabled once per set of tests (which will all have the
same value set).

Note that this error doesn't always occur and had to be identified by
using the ./repeat-test.sh script to run the test multiple times. Prior
to this fix, the test failed on average after 16 runs. After the fix,
the test was able to run through 100 iterations consistently.

BUG=b:213905426
TEST=./repeat-test.sh -s -r 100 -- <DUT> arc.StandardizedKeyboardCopy*

Change-Id: I846c3b494a4e75c4788564042c1f45a24c874316