[Launcher] Fix bubble smoke with sorting enabled
Currently, the bubble smoke test assumes that the system setting app
shows right after showing the bubble launcher. However, when the app
list sorting is enabled along with the bubble launcher, new apps are
always placed at the front. As a result, when the bubble launcher is in
overflow, the setting app button is hidden.

This CL ensures that the bubble smoke test still works with launcher
sorting enabled.

BUG=chromium:1286371
TEST=launcher.BubbleSmoke

Change-Id: Ic0837159b889ff437d6db3215a2e68d232bfae10