Update Tast Test for WordPress
BUG=b:211158139
TEST=tast run <eve> arcappcompat.WordPress.vm_clamshell_mode

WordPress is failing in launch due to a random gmail account
picked up from the test account pool for login using google
account to the app. It is observed that in order to log in
using a google account, WordPress requires a gmail account
to be signed up already. So the test need to use arcappcompat
test account to avoid the flaky failures.

Change-Id: I15df1b641d0013c64a4cdb17f049360a9e7e72c8