lacrosfixt: Reset fixture after failure
SetUp makes the incorrect assumption that a 'prepared' fixture is always
valid. As a consequence, when for example Chrome crashes during a test,
any subsequent tests for the same fixture will fail with a nil pointer
dereference. Fix this by making cleanUp unset the 'prepared' flag, such
that the assumption is correct.

BUG=chromium:1176087
TEST=tast run $DUT policy.PrintingEnabled.lacros*

Change-Id: Ib74448de9fa0f3bf3787eaa9c65cfd63c8821cc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3380494
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Georg Neis <neis@chromium.org>
Commit-Queue: Georg Neis <neis@chromium.org>