faft: When making sure the DUT is on, open ccd
Between tests, we make sure that the DUT is up, in case the test failed
and left it down. Sometimes the DUT can't be booted because the CR50
rebooted during the test, and ccd is locked. To protect against this,
check the CCD level after attaching power before hard reset.

BUG=None
TEST=firmware.BootBatteryCutoff, which is currently broken on leona.
Verified that the fixture recovery booted the machine.

Change-Id: Ife59af2a65c28305b9f99c4f8e2f9ed685ffd38a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3384836
Reviewed-by: Kshitij Shah <tij@google.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>