tast-tests: address frequent timeouts in crostini.BackupRestore
Test fails on arm devices since it is taking longer than 7 mins.
Added extra folders in container.ShrinkDefaultContainer() to reduce size
of backup / restore and make the process faster, and also moved
test to mainlineExpensiveTests where it will have a longer timeout and
run on the minimal set of boards.

BUG=chromium:1257393,chromium:1257379
TEST=crostini.BackupRestore

Change-Id: I6de387ef98197ae2670be7de52a849437a85bec0