uidetection: refactor takeStableScreenshot method
Refactor takeStableScreenshot() to save the current and the
previous screenshots if there is error.

Refactor Location() to save the screenshots when it fails to
find an element on it.

Revise the BasicDetections test to use ImmediateScreenshot
in scenarios where the text cursor is blinking,

BUG=b:213093997
TEST=tast run <DUT> uidetection.BasicDetections

Change-Id: Ib604d98a1f85aa81fd310e40fc18a56bff6ceae9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3380340
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Tested-by: Alvin Jia <alvinjia@google.com>
Commit-Queue: Alvin Jia <alvinjia@google.com>