DO NOT SUBMIT tast-tests: Add lacros variant to ui.DesktopControl
The following tests need Lacros enabled, open tabs and test web content in the browser.
- ui.DesktopControl

Note that it also updates some utils (perfutil, ash.CreateWindows, browser)
to make it work for ash-chrome and lacros-chrome browsers.
The other tests using the same utils could be migrated to Lacros
in follow up CLs when this change looks okay after checked in.

See go/lacros-tast-porting for the porting guidelines.

BUG=None (TBD)
TEST=tast run ui.DesktopControl.lacros
TEST=tast run ui.DesktopControl
TEST=tast run '("group:mainline" && !"dep:lacros" && "name:ui.*")'  # ash-chrome tests

Change-Id: I9799b8f5f1ffb7e47c606f9ff8149bcc132ce0b4