usbprinter: Check usbprinter.Start return before using
If usbprinter.Start returns a non-nil error, it won't return a valid
printer.  We should check the error return before attempting to call
Stop on the returned printer.  I checked the other places where we call
usbprinter.Start and these are the only two that had the incorrect
swapped pattern.

BUG=b:214747966
TEST=tast run printer.*

Change-Id: I7bdaa021de047b30f2d8ed660b7ca882aedade00