tast-tests: Wait for chapsd lock file before cleaning up vault
In test environments, the time difference between mount, unmount
and remove is so short, that chapsd is still writing to the user
cryptohome during the cleanup. Wait for all lock files to
disappear before proceeding to remove the cryptohome vault.

An earlier patch attempted to use pkcs11 terminate to do this.
However, it looks like the call does not terminate chaps' execution.

BUG=b:214879437
TEST=hwssec.KeysetTiedToTPM2 passes

Change-Id: Ia93485c2c76a371538c440c2af3888dc2d4aa565