policy: Test lock screen behavior in QuickUnlockModeAllowlist
Currently this test only tests whether the UI for setting up PIN is
disabled under different settings of the QuickUnlockModeAllowlist
policy. But this policy also controls lock screen behavior, so adding
logic for testing whether the screen can be unlocked by PIN when PIN is
allowed.

Also, we're adding new policies controlling auth methods that will
interact with this policy. The first will be a WebAuthnFactors policy to
control which auth methods can be used for WebAuthn. A user will be
allowed to set a PIN if PIN is included in at least one of these
policies. So there will be upcoming CLs to make this test check that
behaviors of setting a PIN/using a PIN to unlock the screen are correct
given the policy value of these two policies.

BUG=b:214871750
TEST=tast run $DUT policy.QuickUnlockModeAllowlist

Change-Id: I082ed4eb9b79ea7ec4fc91198901344f23224584
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3406367
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>
Reviewed-by: Jana Grill <janagrill@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>