Add launcher folder drag test
Adds a test that verifies a folder item can be dragged within the
current app list page, and to the next / previous page for
paginated app list.

For scrollable app list (i.e. bubble launcher), verifies that the app
list can be scrolled while dragging the folder item, so the folder can
be dropped into initially hidden portion of the apps grid.

BUG=b:190654713
TEST=tast run <DUT> launcher.FolderDragAndDrop*

Change-Id: I19dcff05aba085539e523e625617fb859652ad3b