policy: Add check for fingerprint in QuickUnlockModeAllowlist
The policy QuickUnlockModeAllowlist includes a possible entry
"FINGERPRINT", which on devices with fingerprint hardware enables the
user to set up a fingerprint for authentication. But this test didn't
check anything about fingerprint availability before. Adding a test case
with extra hardware dependency "fingerprint", and in this test case,
check for whether the fingerprint set up UI is present in settings.

BUG=b:214871750
TEST=tast run $DUT policy.QuickUnlockModeAllowlist*

Change-Id: I66c0516e7768f9cf638428f3ec637ecb94bb8a42