security: Refactor entering user session mount namespace
The functionality to set the user session namespace will be
needed in other tests. This CL refactors such functionality
into a helper method so it can be used in other places, such
as a new fixture for guest login that also includes entering
the user session namespace.

BUG=chromium:1157348
TEST=tast run $DUT "*security.U*"

Change-Id: I0d0aed88e5a05e8c3db3256ef5621b40cdbe40dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3383472
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Marcello Salomao <msalomao@google.com>
Commit-Queue: Marcello Salomao <msalomao@google.com>