tast-tests: parse proto in platform.PerfettoChromeProducer
The test performs string matching with the output of `perfetto --query`
to check the number of chromium producers. This is fragile and can be
broken by the format change of `perfetto --query`. Update the test to
parse binary proto from `perfetto --query-raw` so it's unaffected by the
change.

BUG=b:216075225
TEST=`tast run $IP_ADDR platform.PerfettoChromeProducer` passed on hatch

Change-Id: I00c035c4c53aa29e7dd69b698f658178c7364bc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3411645
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Reviewed-by: Cheng-Hao Yang <chenghaoyang@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>