policy: Add WebAuthnFactors test
Add the WebAuthnFactors test that test for whether a PIN can be set and
whether the in session auth dialog is shown when PIN is enabled for
WebAuthn. Check if PIN can be set when either policy includes entry "PIN",
and if WebAuthn is independently controlled by the QuickUnlock policy
only.

BUG=b:214871750
TEST=tast run $DUT policy.WebAuthnFactors*

Change-Id: I7a2af827695af2fdbbeceda09d65ad0738c6b07b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3411963
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Jana Grill <janagrill@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>