dlp: Amends DLP Drag and Drop Tast test to snap windows.
Currently, dlp.DataLeakPreventionRulesListDragdrop is failing on
devices in tablet mode. In tablet mode, snapping a window causes
an overview mode to appear on the non-snapped side, in the
current test no subsequent window is selected and the test
attempts to drag content to the overview window. Hence causing
the test to fail. This CL modifies the flow to snap two windows
(clipboard source and target windows) before attempting to drag
and drop content. This works on both tablet and non-tablet modes.

BUG=chromium:1290248
TEST=tast run <ip> dlp.DataLeakPreventionRulesListDragdrop

Change-Id: I41d60f17d8ff1e397947bfb84ef26cdea0ec49cd