policy: Test print options are available when setting Printers policy
This CL extends the existing test to cover the manual testing steps
linked in b/215675957 and also enables the test for Ash. While extending
the test, I also updated the code to use the `printpreview` to interact
with the print preview dialog where possible.

Due to crbug.com/1290846, interacting with list boxes from Tast doesn't
currently work correctly on Lacros when using `LeftClick`, thus we use
`DoDefault`, which appears to work.

BUG=b:215675957
BUG=chromium:1290846
TEST=tast run <ip> policy.Printers policy.Printers.lacros

Change-Id: I4565d850fbc104ccc8ec3658a0f591fe64977ca5