tast-tests: Add ARC window to ui.WindowArrangementCUJ.*
ui.WindowArrangementCUJ.* will exercise split view resize functionality
with an ARC window that is playing a video.

Note: ui.WindowArrangementCUJ.tablet_mode fails (even without this CL)
because of https://crbug.com/1297355, but I can test this CL on a build
from before that bug existed, and it should work when that bug is fixed.

BUG=b:208290911
TEST=tast run <DUT> ui.WindowArrangementCUJ.*
TEST=tast run <DUT> arc.PIPEnergyAndPower.*
TEST=tast run <DUT> arc.PIPRoundedCornersUnderlay*

Change-Id: I8515f3b748897253a5b8f61308fa9eb842336d80