tast: Add multiple captures case to debugd.PacketCapture
Add multiple packet captures running at the same time test case to
debugd.PacketCapture tast test. Change the part in the test where we
stop the packet capture explicitly for all the test cases to only stop
it when the capture is not expected to terminate itself (due to error or
file size limit).

BUG=b:207359039
TEST=Run the modified tast test by running "tast -verbose run <DUT IP>
debugd.PacketCapture" command.

Change-Id: I159d4effcbec89e830d61043d8441cb84703cdae