arc: Repeat "adb get-state" even after getting error
The "adb get-state" command can fail with the "device unauthorized"
error when it is run just after finishing tests like arc.LibvdaGpuTests
and arc.ChromeCrash.vm_logged_in.

Currently, arc.New() just fails when it gets the unauthorized error.
As a result, we are observing flakiness in several tests when they are
executed in a certain order, despite that the tests themselves may not
have any issues.

Fortunately, the unauthorized state can be temporary and the "adb
get-state" command can succeed after a while in such a case.

In order to reduce the test flakiness, try running the "adb get-state"
command even after getting an error.

BUG=b:197273978
BUG=b:202542613
BUG=b:204917094
BUG=b:216221904
TEST=tast run <DUT> arc.LibvdaGpuTests arc.Downloads.vm

Change-Id: Ic58163ded73bdd331f9bc68dd084e2ed97edcacb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3416931
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>