autoupdate: Make timeouts more modular
Decrease update timeout to 15 minutes.
Create a short timeout for pre-update steps like reading the lbs-relase
file, searching for an image in paygen.json.
Create a shor timeout for post-update steps like checking the new image
version after an update and restoring the original image version.

These steps should be done in a few minutes, but they could use the
whole 20+ minutes allocated for the test.

BUG=None
TEST=tast run -var=updateutil.tlwAddress="<TLW>" \
	 <DUT> autoupdate.BasicNTo*

Change-Id: Idd3dae087c3d44ac9d4bfbf65d49f8f1d07eb42a