[Tast][Launcher] Sorting Smoke Test
Verify that launcher sorting works as expected

Test procedure:

(1) Show the launcher
(2) Move the fake app that should be placed at rear when sorted to the
front. It ensures that apps are out of order before sorting.
(3) Sort with alphabetical order through the item context menu. Verify
that apps are ordered.
(4) Click the undo button. Verify that the app order is recovered.
(5) Sort with alphabetical order again.
(6) Close the launcher then re-show the launcher. Verify that apps
are still ordered.

This CL should be landed after https://crrev.com/c/3419338

BUG=chromium:1227736
TEST=launcher.AppListSortSmoke*

Change-Id: Ifa417ff4305346d9796b19885db6759f58708807