[lacros][tast] Adds a screenshot diff cuj test.
Big visual glitches like crbug.com/1290477 should be be caught during
presubmit/submit testing. This CL adds a new screenshot-diff based test
to validate critical journeys have consistent visual output.

We're adding a single test here for now--one that maximizes the lacros
window and takes screenshots before and after--but are naming the test
generically so that we can add additional CUJs as needed.

BUG=None
TEST=Test passes at head and fails at the broken rev (5ad4b7b13e6cb)

Change-Id: I70cd1af8f9e1811102324276601a067c2f7767c0