tast-test: Add policy.UserFeedbackAllowed test
Test the right behavior of the policy UserFeedbackAllowed
on both Ash and Lacros. This policy allows or denies the
users to send feedback to Google.

It is possible to access the form to send feedback in
two ways: 1) by key combination; 2) from the help menu.
The test should check the policy affects both.

For 1) it is needed to check that the popup appears (or not)
when presing the key combination.

For 2) it is enough to check that the option "Report an issue"
is available (or not) in the Help menu.

BUG=chromium:1288381
TEST=tast run $DUT policy.UserFeedbackAllowed
lacros:tast run $DUT policy.UserFeedbackAllowed.lacros
both:tast run $DUT policy.UserFeedbackAllowed*

Change-Id: I012f46991a06f617c4ba23e9d51d81ff9d1a7bb5