tast-tests: test custom name/icon from policy
The WebAppInstallForceList policy can specify custom icons
and names for the installed apps.

Previously, it was not possible for those values to be
omitted. This CL modified policy/defs.go to make all
non-required policy fields (all but "url") omittable.

The test itself is changed to use the app settings page
for easier title-and-icon-checking without animation delay and
refresh issues.

Also, more test cases were added to cover PWAs with manifests,
web sites without one, and placeholder apps where the
web site could not be reached.

BUG=b:215521658
TEST=tast run <DUT> policy.WebAppInstallForceList

Change-Id: Id9b09424f54e314464bf29f294a287bcfb06125f