policy: Extend the tast test for AutofillCreditCardEnabled
Extending the test to add a credit card in Chrome's settings and if the
policy is enabled, to actually autofill the saved credit card details
into a test page having a credit card autofill enabled form. Autofilling
credit card only works in HTTPS and thus also a HTTPS server with a
private key and a certificate are needed.

BUG=chromium:1288277
TEST=tast run <ip> policy.AutofillCreditCardEnabled

Change-Id: I6fec88f7388c71a7636c3fe9e3c3d5c1e68da034