tast-tests: Adjust divider drag in ui.WindowArrangementCUJ.*
BEFORE: The tests ui.WindowArrangementCUJ.* involve dragging the divider
all the way right, then to a position slightly left of the center, and
finally back to the center. This is different from the divider drag in
ui.SplitViewResizePerf*, but there is no apparent motivation for the
difference. If the divider drag in ui.WindowArrangementCUJ.* fails, then
an error message reads "failed to drag divider slightly right, all the
way left, and back to center," which does not accurately describe any
part of any test that I know of.

AFTER: The tests ui.WindowArrangementCUJ.* involve dragging the divider
slightly left, all the way right, and back to the center. This is
consistent with the divider drag in ui.SplitViewResizePerf*. If the
divider drag in ui.WindowArrangementCUJ.* fails, then an error message
reads "failed to drag divider slightly left, all the way right, and back
to center."

BUG=chromium:1087182
TEST=tast run <DUT> ui.WindowArrangementCUJ.*

Change-Id: Ife8ae8a53e72ee5def56953c0f8031f0c10b00a6