[Tast][Launcher] Fix flakiness of BubbleSmoke test
This CL does the following things:

(1) The setting app should be searched from the scrollable apps grid instead
of the app list bubble view in order to rule out the possibility that
the setting app could exist in the continue section. It fixs the error that "multiple nodes matched ... ".

(2) Wait for the launcher to become closed after switching to the
clamshell mode. It fixes the error that "Failed to wait for location changes: context
deadline exceeded".

BUG=chromium:1289371
TEST=launcher.BubbleSmoke*

Change-Id: I632118bc1033c7432833769289bac97bbb9f70a3