arcappcompat: Adjust trackpad zoom to use relative coordinates
Adjust the trackpad zoom to use coordinates relative to the size of the
trackpad, rather than the fixed amount. This allows the full trackpad to
be used, rather than a small portion. Additionally, by basing the zoom
on the size of the trackpad, the test becomes more consistent for all
boards.

BUG=b:216974685
TEST=push apk && tast run $DUT arc.standardized*

Change-Id: I46f0174c14f90112072da7779d702a6df5a6abe5