arc: Stop falling back to localhost when emulator is found in connectADB
Currently, connectADB tries to fall back to connecting to localhost:5555
if 1) an emulator is found, but 2) WaitForState reaches its timeout.

However, when this happens, it does not report why / how WaitForState
reaches its timeout. When WaitForState reaches its timeout, the timeout
for connectADB is also reached, and the entire test fails just saying
"failed connecting to ADB: context deadline exceeded" without any
detailed report on WaitForState failure reasons.

In order to get more meaningful logs, we need to report the reasons of
WaitForState failures by either logging or returning errors.

Here, just returning errors should be sufficient. This is because
crrev.com/c/3416931 made WaitForState fail only when it reaches its
timeout (= connectADB's timeout), and thus there is no point in falling
back to localhost on WaitForState failures now.

As for the justification of crrev.com/c/3416931, recent test results
indicate that it actually stabilized arc.New*, and no or only a few new
failures were introduced by it. However, in order to confirm that, we
need logs, and this CL will help with it.

With this CL, fallback to localhost is performed only when no emulator
is found by WaitForDevice (in line with the original implementation by
crrev.com/c/2446602).

BUG=b:197273978
BUG=b:202542613
BUG=b:204917094
BUG=b:216221904
TEST=tast run <DUT> arc.LibvdaGpuTests arc.Downloads.vm
TEST=Confirm that the failure reason is properly reported when an
     emulator is found but WaitForState reaches its timeout (e.g. by
     setting a very short timeout value).

Change-Id: Ic8748a66188e9a97eb34b66187f4beba4c8f3e83