tast-tests: Tune traffic control params for arc.AppLoadingPerf on ARM
Original tc-tbf parameters was only tuned for x86/_64 boards. They
need to be re-tuned for ARM to match iperf3 bandwidth and Play Store
game download tests performed by the ARCVM network team on kukui vs.
kukui-arc-r. This needs to be verified on more ARM boards once we
have Crosbolt data with this CL.

BUG=b:215621884
TEST=tast run $DUT arc.AppLoadingPerf* on kukui vs. kukui-arc-r

Change-Id: I3ac5d169907092f97d0064df19aa4d77dba6c652