arc: chromeproxy: Define optional JSON fields of struct types as ptrs
The "omitempty" option of golang's json package is not directly
applicable to struct types. Due to this, currently, JSON encoding of
chromeproxy.ProxyConfig always contains redundant parameters.

Such redundant parameters are not allowed by API, and not expected by
Chrome. As a result, arc.SettingsBridge(.vm) Tast tests are failing with
Chrome DCHECK failures on boards with dcheck_always_on = true.

In order to apply the "omitempty" option to redundant parameters, define
optional JSON fields of struct types as pointers.

BUG=b:217100326
TEST=Build & deploy Chrome with dcheck_always_on = true, and
     $ tast run <DUT> arc.SettingsBridge(.vm)

Change-Id: I0267ae095ec1f30ac0e8a4b357951ed6418124e0