policy: Revisit UserAvatarImage logic
For the UserAvatarImage policy test, the check for animated account
images is flaky.

For the unset test case, this CL changes the verification logic to rely
on more robust check: compare the image source of the image preview and
the selected avatar image.

For the non_empty test case, although the image is not cropped to a
circle, the bounds of the taken screenshot lie completely inside the
circle, so 100% red pixels.

BUG=chromium:1257963
TEST=tast run <IP> policy.UserAvatarImage

Change-Id: I0d2eaf5f84fb2656d17da954099ae3685e5f2f5f