policy: Revisit UserAvatarImage logic
Before this CL, the UserAvatarImage policy test had many bugs:

1- Due to UI changes, the preview for the policy-provided image has
different dimensions than normal avatar images in some devices. That
resulted in a bug where the preview screenshot has inconsistent number
of red pixels. This CL fixes that by recalculating the preview
dimensions before taking the screenshot.

2- The check for animated account images is flaky. This CL removes the
Poll logic to rely on more robust check: compare the image source of
the image preview and the selected avatar image.

3- The unset use case had an uncaught bug where the chosen avatar
wasn't clicked on because the avatar button wasn't visible. The new
logic introduced in (2) above exposes this bug as we now compare the
source of both images. This CL fixes that by using MakeVisible() before
clicking on the node.

BUG=chromium:1257963
TEST=tast run <IP> policy.UserAvatarImage

Change-Id: I0d2eaf5f84fb2656d17da954099ae3685e5f2f5f