arcappcompat: Update Tast Test for Pandora
BUG=b:193837758
TEST=tast run <helios> arcappcompat.Pandora.vm_tablet_mode
tast run <krane> arcappcompat.Pandora.tablet_mode

When the device is in tablet mode, Pandora is failing in
launch since the login button is covered by a virtual
keyboard. Handled it in a code. Also observed that on
some devices, a new dialog box for saving the password
appears. Added a code to handle it as well. This helps
to fix the flaky failures.

Change-Id: I92be2023a7166a404dd0766610b542da99bb0245