arcappcompat: Adjust touchscreen to account for rotation
Adjust the touchscreen to account for rotation of the display when
applying coordinates. Without this, the coordinates that are returned
for elements cannot be properly converted to touchscreen coordinates on
devices which are rotated. In order to generalize this process for
the Standardized suite, a new type was added to enforce usage of the
configured touchscreen instance.

BUG=b:217464962
TEST=push apk && tast run $DUT arc.StandardizedTouch*

Change-Id: Ie4a602790ceb9d38afa3fd165e88376fa2200324