scanapp: remove Canon MF741C/743C scanner descriptor
This descriptor was used for local development and was committed to
serve as an example for future scanner descriptors. This scanner is not
present in the lab, and now that there are other scanner descriptors in
the tree there's no reason to keep this one around.

BUG=b:195064784
TEST=CQ passes

Change-Id: I633ace64079da766e3c73becbe75a9e649bc5db6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3433025
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Reviewed-by: Matt Stark <msta@google.com>
Tested-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Paul Moy <pmoy@chromium.org>