DO NOT SUBMIT (on hold) refactor: Register Closer for Browser.Close
This will register a closer function to be called in Browser.Close
to clean up browser and resources for underlying browser.

With this,
br.Close replaces closeBrowser closure returned from browserfixt.SetUp
br.Close is l.Close for lacros-chrome
br.Close does nothing for ash-chrome as a browser now, but could be updated when needed
cr.Close co-exists for ash-chrome along with br.Close

Also fixed a potential issue to avoid a runtime error
when (Lacros).Close is called on already closed session.

See go/lacros-tast-api-comparison for the discussion.
See go/lacros-tast-porting for the porting guidelines.

BUG=chromium:1254152
TEST=tast run example.Browser*

Change-Id: I1d2f5d6203c267df450c7cf146cf0e0b631ac33e