playstore: Allow the app to install until progress bar is gone
BUG=b:209036276
TEST=tast run <helios> arcappcompat.Asphalt.vm_clamshell_mode

In the current version of Playstore.go, it does not wait for the
installation progress bar to go away. This CL waits until the
installation progress bar is gone and then clicks on the open
button. This will help to fix the flaky installation failures.

Change-Id: I37fe94ceac693cb27bd7ea249ec5e065a08556fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3441088
Tested-by: Meenakshi Thiyagarajan <mthiyagarajan@google.com>
Reviewed-by: David Welling <davidwelling@google.com>
Commit-Queue: Meenakshi Thiyagarajan <mthiyagarajan@google.com>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>