arcappcompat: Allow the app to install until progress bar is gone
BUG=b:217964717
TEST=tast run <kevin> arcappcompat.BoostedProductivity.clamshell_mode
tast run <helios> arcappcompat.Asphalt.clamshell_mode

In the current version of Playstore.go, it does not wait for the
installation progress bar to go away. This CL waits until the
installation progress bar is gone and then clicks on the open
button. This will help fix the installation failures.

Change-Id: I37fe94ceac693cb27bd7ea249ec5e065a08556fb