arcappcompat: Allow the app to install until progress bar is gone
BUG=b:217964717, b:209036276
TEST=tast run <kevin> arcappcompat.Snapchat.clamshell_mode
tast run <helios> arcappcompat.Asphalt.vm_clamshell_mode

In the current version of Playstore.go, it does not wait for the
installation progress bar to go away. This CL waits until the
installation progress bar is gone and then clicks on the open
button. Also added an extra check for "Want to link your PayPal
account" and “Complete account setup” text messages to verify
if the UI is enabled for them.  So the code inside the  "Want
to link your PayPal account" and “Complete account setup” block
will be executed only if they are enabled. This will help fix
the flaky installation failures.

Change-Id: I37fe94ceac693cb27bd7ea249ec5e065a08556fb