tast-tests: Fix setting white wallpaper on small screen
The solid white wallpaper is at the end of the wallpaper list so when
the screen height is too small the wallpaper isn’t visible.
We need to scroll down to make sure the white wallpaper is visible.

This CL also extracts the logic to the wm helper package.

BUG=b:217812426
TEST=arc.CompatSnap
TEST=arc.ResizeLock

Change-Id: Ibbc64eeda54da4fb2c6d49a56fd966631a4efec8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3439440
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Commit-Queue: Toshiki Kikuchi <toshikikikuchi@chromium.org>