typec: Don't fail the test on CCD keepalive setting failure
On some boards, servod occasionally doesn't get the servod interfaces
back on at the conclusion of the test (when we enable CC). So, the
cleanup call to enable CCD keepalive occasionally fails.

Change the "Error" print to a "Log" print when this happens; we
shouldn't fail the test because servod was unable to restart the
interfaces as part of post-test cleanup.

BUG=b:190578402
TEST=typec.Basic passes 100/100 times on volteer.

Change-Id: I636caf802ca7d8c2774ab86b828f32be6beb06f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3448335
Tested-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Prashant Malani <pmalani@chromium.org>