camera: Use the mapped volume keys instead in CCAUIVolumeShutter
After https://crrev.com/c/3343820, CCAUIVolumeShutter might try to press
"volumedown" or "volumeup" key on keyboard which does not have physical
volume keys, and cause the test to fail.

This CL changes to use the mapped volume keys so that the test can pass
on devices with keyboard that don't have physical volume keys but have
combination keys for volume control.

BUG=b:218436978
TEST=tast run [DUT] camera.CCAUIVolumeShutter

Change-Id: Iadf82e7776016a2ac4a23c23008c167b23ed6357