samba: Add filemanager.SMBFileOperations copy test
This adds the necessary scaffolding for testing all the file operations
from and to a Samba share. This initially implements the copy operation
but leaves it generic enough to implement the remaining operations.

In addition to the above it updates the fixture to ensure the smb.conf
file forces itself as chronos and subsequently chown's the guestshare to
ensure the user can write to try.

BUG=chromium:1295640
TEST=tast run DUT filemanager.SMBFileOperations*

Change-Id: I8e437de3171b17438df8227de2b95de49469ba99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3450252
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Pranav Batra <batrapranav@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>