arc: Use Files app to test volume sharing in arc.RemovableMedia
We are going to restrict file access to ARC volumes so that files can be
served by ArcVolumeProvider only when they are opened by the Files app.

Currently, arc.RemovableMedia directly calls adb's content method to
read a file using ArcVolumeProvider. Since it is doomed to fail after
the restriction, we need to change the test's implementation.

This CL modifies the arc.RemovableMedia test so that it uses the CrOS
Files app to open a file in a removable media volume with an Android
app. This is compatible with the planned restriction, and also close to
the real world use cases.

This CL also adds a check for file sharing from Android to Chrome OS,
which is not checked by the current implementation.

Since these checks can be done in almost the same way as arc.MyFiles, a
new utility function storage.TestVolumeSharing is defined and used by
both tests.

Also, the test is demoted to informational as the new test flows are not
verified well.

This CL depends on a Chromium CL crrev.com/c/3450055.

BUG=b:179656745
TEST=tast run <DUT> arc.RemovableMedia* arc.MyFiles* \
     arc.DownloadsFolder* arc.Drivefs* arc.Fsp* arc.MTP*

Change-Id: I29995d2ee8a8cedab3663623e8524bbc74fdd1a4