u2fd: Add WebauthnUsingPassword integration test
Add WebauthnUsingPassword integration test to test the behavior of
WebAuthn using password. Refactor the common codes with WebauthnUsingPin
to the utils package.

Also, slightly change the logic to determine whether MakeCredential and
GetAssertion succeeded, and increase the timeout since the new test
often takes around the current timeout to run.

BUG=b:218790804
TEST=tast run $DUT u2fd.WebauthnUsing*

Change-Id: I355c8812afdfcb075b99d2156622eeb6ed19ced1