tast-tests: Update the playbilling test apk to use the latest icon
The playbililng test icon was incorrectly using the web share target
icon in the APK. This was not matching with the icon loaded from disk
and causing an overlay to occur. Updated and rolled out a new APK with
the correct icon.

This also updates the port number as the new bundled apk was generated
using the port 8080. This was mainly due to the fact port 80 cannot
be bound to on corp desktops so switched to a port that will enable
future testing (initial APK was initially not served on a corp machine).

BUG=chromium:1290626
TEST=tast run DUT appsplatform.CheckSkuPurchase*

Change-Id: Icf16a4fe7c2d5aaedb584fe03b0f57895114cf52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3447741
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>