u2fd: Check UI correctness in WebAuthn tests
Currently tast doesn't interact with the in session auth dialog's UI in
WebAuthn. As soon as it appears, we type the credential using the
keyboard and pressed enter. We might want to check if the correct UI
components are present. For example, if PIN auto-submit is enabled, the
UI should show the LoginPinInputView and not normal textfield. Add
checks for the correct UI component in WebAuthnUsing* tests.

BUG=b:218790804
TEST=tast run $DUT u2fd.WebauthnUsing*

Change-Id: I9ee76273f977306fc8acab6641c0cccd7ca60bbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3447746
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>