tast-tests: Stabilize accountmanager.AddProfile*
The tests are verifying creation of a new Chrome Profile. Currently the
focused Chrome window is used to find the Profile which was just
created. This makes the test very flaky, because the newly opened Chrome
window is not always being focused.

To find the new Chrome window out of two windows present:
- Open chrome://version in the first Chrome window.
- Create a new Chrome Profile.
- Find the window which doesn't contain "About Version" in its name.

BUG=chromium:1268351
TEST=tast run <ip> accountmanager.AddProfile*

Change-Id: I32da5d4abc8c7ba1708ba2a67ab609f5413f5fed