Finch smoke test: fix failures in fetching production seed
The test occasionally fails when trying to find production Finch seed
values in Local State. It can take some time to fetch the seed from the
server and for it to show up in Local State, so poll for it to appear to
account for this.

BUG=b:202325793
TEST=./repeat-test.sh -s -r 100 -- localhost:2223 ui.ChromeVariationsSmoke

Change-Id: Ice1a771819e9d48e3a70bdc93462cbdff820618f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3457171
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>