tast: Do not validate unexpected force installed apps
The test helper is used to ensure that force installed apps are in the
policy that the test is going to wait for for installation. The test
currently fails if there are apps that were not expected. This
additional check is unnecessary since anyone modifying the org or the
root org that the test user is part of can cause additional apps to
appear for the test. It is sufficient to only check that the apps that
are expected are preset in the policy for the test to succeed.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm
TEST=/fast_build.sh -T

Change-Id: I89e6871a8af04072d4c9dc92a92b42507b8d5fd2