uidetection: revise the log for stable screenshot
The coordinates in the previous log when two images are not equal
are the absolute coordinates in the whole screenshot. On the other
hand, the saved images are the cropped images if relative locations
are used.

Revised to make the coordinates consistent with the saved image
by using the relative position in the bounds.

BUG=None
TEST=tast run <DUT> uidetection.Basic*

Change-Id: Ib9e47fa625a6432ba36ac5494efe2db340dd1574
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3460623
Tested-by: Alvin Jia <alvinjia@google.com>
Reviewed-by: Matt Stark <msta@google.com>
Commit-Queue: Alvin Jia <alvinjia@google.com>