tast-tests: enable DFS when SpectrumManagement presented in AP config
The SpectrumManagement hostapd config setting appeared to be meant to enable DFS, but it did not previously turn on ieee80211h, the hostapd setting for DFS. Now when SpectrumManagement is enabled, so is ieee80211h. The SimpleConnect tests using channels where DFS should be enabled in the US now turn on SpectrumManagement as well.

BUG=b:211517577
TEST=tast run -var=router=$ROUTER -var=pcap=$PCAP $DUT wifi.SimpleConnect.*

Change-Id: I770f43c0dea344c696e047a578296bb34bc14e0b